pull out awg interfaces into new file, allow for strings as parameter values, make repeated element work by forcing value to be int rather than numpy int64, add units and labels to repeated element setpoints
